rename: ~api/members -> ~api/member
