+++ src/apis/member/useUser.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import { fetchUser } from '~apis/members/fetchUser';
+import { fetchUser } from '~apis/member/fetchUser';
 
 export const useUser = () => {
   const { data: userInfo } = useSuspenseQuery({
